services/browser-agent/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
anthropic>=0.52.0,<1.0.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
browserbase>=1.0.0,<2.0.0
stagehand>=1.0.0,<2.0.0
playwright>=1.49.0,<2.0.0
httpx>=0.27.0,<1.0.0
azure-storage-blob>=12.24.0,<13.0.0
azure-identity>=1.19.0,<2.0.0
